RZO-442 ~ We can edit email field now.
